Revert "Make autowall value an int"
This reverts commit ba83c66e5140671c46a30a8cd3112c698084177b.

+++ src/settings.cpp
@@ -260,7 +260,7 @@ void Settings::LoadSettings(const char* filename)
 	GetFloat(settings["Aimbot"]["Smooth"]["Salting"]["percentage"], &Settings::Aimbot::Smooth::Salting::percentage);
 	GetBool(settings["Aimbot"]["AutoAim"]["enabled"], &Settings::Aimbot::AutoAim::enabled);
 	GetBool(settings["Aimbot"]["AutoWall"]["enabled"], &Settings::Aimbot::AutoWall::enabled);
-	GetInt(settings["Aimbot"]["AutoWall"]["value"], &Settings::Aimbot::AutoWall::value);
+	GetFloat(settings["Aimbot"]["AutoWall"]["value"], &Settings::Aimbot::AutoWall::value);
 	Settings::Aimbot::AutoWall::bones.clear();
 	for (Json::ValueIterator itr = settings["Aimbot"]["AutoWall"]["bones"].begin(); itr != settings["Aimbot"]["AutoWall"]["bones"].end(); itr++)
 		Settings::Aimbot::AutoWall::bones.push_back((Hitbox) settings["Aimbot"]["AutoWall"]["bones"][itr.index()].asInt());
